Extact retrofit setup code to new classes
