+++ vimeo-networking/src/main/java/com/vimeo/networking/Vimeo.java
@@ -32,7 +32,7 @@ public final class Vimeo {
 
     public static final String VIMEO_BASE_URL_STRING = "https://api.vimeo.com/";
 
-    static final String API_VERSION = "3.3.3";
+    public static final String API_VERSION = "3.3.3";
 
     // Global Constants
     public static final int NOT_FOUND = -1;
